code refactoring

+++ nagp-microservices-devops/consumer-service/src/main/java/com/consumer/model/Feedback.java
@@ -12,7 +12,7 @@ public class Feedback {
 	private Long id;
 	private Long bookingId;
 	private Long serviceExpertId;
-	private int ratingValue;
+	private double ratingValue;
 	private String description;
 	private LocalDateTime createdAt;
 }
